// typings/GroupedComboboxProps.d.ts
export type OptionsSourceAssociationCustomContentTypeEnum = "yes" | "listItem" | "no";